==> RecSysProj/lab_ts/05 ensemble als+i2i.ipynb <==
%pylab inline
import pandas as pd
import numpy as np
import seaborn as sns
import os
import sys

df_als = pd.read_csv("/data/home/taras.svirsky/lab12/res/res_als_r25_ni15.csv")
print df_als.shape
df_als.head(2)

df_i2i = pd.read_csv("/data/home/taras.svirsky/lab12/res/full_test_bp_i2i_top100.csv")
print df_i2i.shape
df_i2i.head(2)

df_als = df_als.rename(columns={"rating":"als_rating"})
df_i2i = df_i2i.rename(columns={"predicted_i2i_rating":"i2i_rating"})

df = df_i2i.merge(df_als, on=["userId","movieId"], how='left')

idx = df["als_rating"].isnull()
df.loc[idx, "als_rating"] = df.loc[idx, "bp_predicted_rating"]

del df_als
del df_i2i

df_cat = pd.read_csv("/data/home/taras.svirsky/lab12/res/test_subm_rating_counts.csv")

df = df.merge(df_cat, on=["userId","movieId"], how='left')

df['n_user_ratings'] = df['n_user_ratings'].fillna(0)
df['n_item_ratings'] = df['n_item_ratings'].fillna(0)

sum(df.isnull())

assert df.shape[0] == 10531564

del df_cat

# combine all

def cut_dot5to5(x):
    if x < .5:
        return .5
    if x > 5:
        return 5.0
    return x

idx_big = (df.n_user_ratings > 50) & (df.n_item_ratings>100)
idx_med = (df.n_user_ratings > 5) & (df.n_user_ratings <= 50) & (df.n_item_ratings>100)
idx_small = (~idx_big) & (~idx_med)

print sum(idx_big), sum(idx_med), sum(idx_small)
n = sum(idx_big) + sum(idx_med) + sum(idx_small)
assert n == 10531564
print 100.*sum(idx_big)/n, 100.*sum(idx_med)/n, 100.*sum(idx_small)/n


# v 351
#df["rating"] = np.where((df.n_item_ratings>100) &  (df.n_user_ratings > 5),
#                        0.88 * df["als_rating"] + 0.15 * df["i2i_rating"],
#                        0.38 * df["als_rating"] + 0.62 * df["i2i_rating"],
#                       )

# v 8010
df["rating"] = df["bp_predicted_rating"]
df['i2i_x_als'] = df['als_rating']*df['i2i_rating']

df.loc[idx_big,"rating"] =  -0.35258 * df.loc[idx_big, "bp_predicted_rating"] \
                          + 1.23212959 * df.loc[idx_big, "als_rating"] \
                          + 0.15828964 * df.loc[idx_big, "i2i_rating"]

df.loc[idx_med,"rating"] =  0.933 * df.loc[idx_med, "als_rating"] \
                          + 0.211 * df.loc[idx_med, "i2i_rating"] \
                          -0.0292 * df.loc[idx_med, "i2i_x_als"]


df.loc[idx_small,"rating"] =  0.483 * df.loc[idx_small, "als_rating"] \
                             + 0.65 * df.loc[idx_small, "i2i_rating"] \
                             -0.034 * df.loc[idx_small, "i2i_x_als"]



res = df[["userId", "movieId", "rating"]].sort_values(by=["userId", "movieId"])

res.rating.describe()

res['rating'] = res['rating'].map(cut_dot5to5)

res.rating.describe()

res.to_csv("/data/home/taras.svirsky/lab12/res/full_test_v8010.csv", index=False)
