==> RecSysProj/lab_ts/21 find best ensembling.ipynb <==
import pandas as pd
import numpy as np
%pylab inline
import seaborn

df = pd.read_csv("/data/home/taras.svirsky/lab12/data/stest_pivot4ensemble.csv")

df.shape

#sum(df.isnull())

def get_rmse(sdf, column):
    e = sdf['rating'] - sdf[column]
    return np.sqrt(np.mean(e*e))

for f in ['bp_rating', 'predicted_i2i_rating', 'predicted_u2u_rating', 'predicted_als_rating']:
    print f, get_rmse(df, f)

df['pred'] = 0.75 * df["predicted_als_rating"] + 0.25 * df["predicted_i2i_rating"]

get_rmse(df,'pred')

recs = []
for j in range(20):
    for i in range(20):
        r1 = 0.05*i
        r2 = 0.05*j
        r3 = 1 - r1 - r2
        if r3 < 0:
            continue
        df['pred'] = r1 * df["predicted_als_rating"] + r2 * df["predicted_u2u_rating"] + r3 * df["predicted_i2i_rating"]
        recs.append((r1,r2,r3, get_rmse(df,'pred')))
        
sdf = pd.DataFrame.from_records(recs, columns=["als", "u2u", "i2i", "rmse"])        

sdf.sort_values(by="rmse").head()

# looks like we have overfit on u2u, because it gives 0.86 rmse on submission

# so I think to use second row

#als 0.35 u2u 0.50 i2i 0.15
#als 0.6  u2u 0.3  i2i 0.1
#als 0.50 u2u 0.35 i2i 0.15

# just for fun

from sklearn.cross_validation import cross_val_score, train_test_split

from sklearn.cross_validation import KFold
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error

lr = LinearRegression(fit_intercept=False)

idx_big = (df.n_user_ratings > 50) & (df.n_item_ratings>100)
idx_med = (df.n_user_ratings > 5) & (df.n_user_ratings <= 50) & (df.n_item_ratings>100)

idx = (~idx_med) & (~idx_big)

df['i2i_x_als'] = df['predicted_als_rating']*df['predicted_i2i_rating']

X = df[idx][["predicted_als_rating", "predicted_i2i_rating", "i2i_x_als"]]

y = df[idx]['rating']

cv = KFold(len(df[idx]), n_folds=10, shuffle=True, random_state=17)

z=cross_val_score(lr, X, y, cv=cv, scoring='neg_mean_squared_error', n_jobs=4)

-np.mean(z)

X_train, X_test, y_train, y_test = train_test_split(X,y, train_size=0.8, random_state=434)

lr.fit(X_train, y_train)

lr.coef_

mean_squared_error(y_test, lr.predict(X_test))

lr.intercept_

# big -0.35258075,  1.23212959,  0.15828964
# med 

array([ 0.48307975,  0.65027222, -0.03491632])
